Add files via upload
Sample dataset of 4000 customers across five cities on their online spending and other details are provided to analyse the sample data and report key findings to the strategy team of Mitron Bank. This analysis is expected to guide them in tailoring the credit cards to customer needs and market trends.